requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
pandas>=2.0.0
python-dateutil>=2.8.0
